Add Task.java Move Task and subclasses to Task.java
