+++ src/main/java/Yale.java
@@ -20,8 +20,6 @@ public class Yale {
     private static final Pattern EVENT_REGEX = Pattern.compile("event (.+) /from (.+) /to (.+)");
     private static final Pattern DELETE_REGEX = Pattern.compile("delete (\\d+)");
 
-    private static final DateTimeFormatter DATE_FORMAT = DateTimeFormatter.ofPattern("MMM dd yyyy");
-
     private static final String[][] COMMANDS = {
             {"mark", "mark [id]"},
             {"unmark", "unmark [id]"},
@@ -63,11 +61,11 @@ public class Yale {
             } else if ((m = UNMARK_REGEX.matcher(msg)).matches()) {
                 markDone(Integer.parseInt(m.group(1)), false);
             } else if ((m = TODO_REGEX.matcher(msg)).matches()) {
-                addTask(new ToDo(m.group(1)));
+                addTask(new Task.ToDo(m.group(1)));
             } else if ((m = DEADLINE_REGEX.matcher(msg)).matches()) {
-                addTask(new Deadline(m.group(1), m.group(2)));
+                addTask(new Task.Deadline(m.group(1), m.group(2)));
             } else if ((m = EVENT_REGEX.matcher(msg)).matches()) {
-                addTask(new Event(m.group(1), m.group(2), m.group(3)));
+                addTask(new Task.Event(m.group(1), m.group(2), m.group(3)));
             } else if ((m = DELETE_REGEX.matcher(msg)).matches()) {
                 deleteTask(Integer.parseInt(m.group(1)));
             } else {
@@ -137,89 +135,6 @@ public class Yale {
         }
         ui.printError("Sorry, I don't know what that command means.");
         ui.endOutput();
-    }
-
-    public static LocalDate tryParseDate(String dateStr) {
-        try {
-            return LocalDate.parse(dateStr);
-        } catch (DateTimeParseException e) {
-            return null;
-        }
-    }
-
-    private static abstract class Task {
-        protected final String name;
-        protected boolean isDone = false;
-        public Task(String name) {
-            this.name = name;
-        }
-        public void setDone(boolean done) {
-            this.isDone = done;
-        }
-        public String toString() {
-            return "[%s] %s".formatted(isDone ? "X" : " ", name);
-        }
-        protected abstract String[] getParams();
-        public String toCsv() {
-            return String.join("\0", getParams());
-        }
-        public static Task fromCsv(String csvString) {
-            String[] tokens = csvString.split("\0");
-            return switch (tokens.length) {
-                case 2 -> new ToDo(tokens[1]);
-                case 3 -> new Deadline(tokens[1], tokens[2]);
-                case 4 -> new Event(tokens[1], tokens[2], tokens[3]);
-                default -> null;
-            };
-        }
-    }
-
-    private static class ToDo extends Task {
-        public ToDo(String name) {
-            super(name);
-        }
-        public String toString() {
-            return "[T]%s".formatted(super.toString());
-        }
-        protected String[] getParams() {
-            return new String[] {isDone ? "X" : " ", name};
-        }
-    }
-
-    private static class Deadline extends Task {
-        private final String deadlineStr;
-        private final LocalDate deadlineDate;
-        public Deadline(String name, String deadline) {
-            super(name);
-            this.deadlineDate = tryParseDate(deadline);
-            this.deadlineStr = (this.deadlineDate == null) ? deadline : this.deadlineDate.format(DATE_FORMAT);
-        }
-        public String toString() {
-            return "[D]%s (by: %s)".formatted(super.toString(), deadlineStr);
-        }
-        protected String[] getParams() {
-            return new String[] {isDone ? "X" : " ", name, deadlineStr};
-        }
-    }
-
-    private static class Event extends Task {
-        private final String startStr;
-        private final String endStr;
-        private final LocalDate startDate;
-        private final LocalDate endDate;
-        public Event(String name, String start, String end) {
-            super(name);
-            this.startDate = tryParseDate(start);
-            this.startStr = (this.startDate == null) ? start : this.startDate.format(DATE_FORMAT);
-            this.endDate = tryParseDate(end);
-            this.endStr = (this.endDate == null) ? end : this.endDate.format(DATE_FORMAT);
-        }
-        public String toString() {
-            return "[E]%s (from: %s, to: %s)".formatted(super.toString(), startStr, endStr);
-        }
-        protected String[] getParams() {
-            return new String[]{isDone ? "X" : " ", name, startStr, endStr};
-        }
     }
 
     private void markDone(int id, boolean done) {
